@@ -1,3 +1,3 @@
 module mynews
 
-go 1.21
+go 1.25
